Gui: move bootstrap callbacks into ApplicationPy
Move showMainWindow, exec_loop, setupWithoutGUI, and embedToWindow from the FreeCADGui entry module into Gui::ApplicationPy.

This keeps FreeCADGuiPy.cpp focused on module initialization and gives generated FreeCADGui module bindings a stable C++ callback owner for both bootstrap and full-GUI exports.
